Adding files used to create current cards
The zip file contains all of the images for a variety of cards to choose
from. I included the inkscape files for the cards I ended up using, as
well as the gimp files I used to generate the png files used in the code
repo. The gimp files are after resizing and adding appropriate empty
space so that the sprite sheet could be used properly.